Restructure source
