+++ src/main/java/bun/task/Event.java
@@ -1,4 +1,4 @@
-package bun.ui;
+package bun.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
