+++ src/main/java/bun/task/Task.java
@@ -1,4 +1,4 @@
-package bun.ui;
+package bun.task;
 
 public class Task {
