+++ src/main/java/bun/task/ToDo.java
@@ -1,4 +1,4 @@
-package bun.ui;
+package bun.task;
 
 public class ToDo extends Task {
     public ToDo(String description) {
